refactor BUConfig.register to ConfigUtil
