+++ src/main/java/com/github/mkram17/bazaarutils/config/util/ConfigUtil.java
@@ -2,25 +2,29 @@ package com.github.mkram17.bazaarutils.config.util;
 
 import com.github.mkram17.bazaarutils.BazaarUtils;
 import com.github.mkram17.bazaarutils.config.BUConfig;
+import com.github.mkram17.bazaarutils.config.patcher.ConfigPatches;
 import com.github.mkram17.bazaarutils.utils.Util;
+import com.google.gson.JsonObject;
 import com.teamresourceful.resourcefulconfig.api.client.ResourcefulConfigScreen;
+import com.teamresourceful.resourcefulconfig.api.loader.Configurator;
+import com.teamresourceful.resourcefulconfig.api.types.ResourcefulConfig;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.widget.ClickableWidget;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.function.UnaryOperator;
 
 import static com.github.mkram17.bazaarutils.BazaarUtils.CONFIGURATOR;
 
 
 public class ConfigUtil {
 
+    public static final Map<Integer, UnaryOperator<JsonObject>> PATCHES = ConfigPatches.loadPatches();
+    public static final int VERSION = 1;
     private static boolean configSaveScheduled = false;
-
-    public static void registerConfig() {
-        BUConfig.register(CONFIGURATOR);
-    }
 
     public static Screen createGUI(Screen parent) {
         return ResourcefulConfigScreen.make(BazaarUtils.config)
@@ -49,5 +53,27 @@ public class ConfigUtil {
         List<ClickableWidget> widgets = new ArrayList<>();
         //automatically added using @RegisterWidget annotation
         return widgets;
+    }
+
+    public static ResourcefulConfig register(Configurator configurator) {
+        if (PATCHES.size() + 1 != VERSION) {
+            throw new IllegalStateException(
+                    "BUConfig VERSION (" + VERSION + ") is out of sync with patch count " +
+                            "— expected VERSION = " + (PATCHES.size() + 1)
+            );
+        }
+
+        configurator.register(BUConfig.class, event ->
+                PATCHES.forEach((version, patch) ->
+                        event.register(version, json -> {
+                            Util.logMessage("Applying patch " + version);
+                            JsonObject result = patch.apply(json);
+                            Util.logMessage("[BUConfig] Patch " + version + " applied successfully");
+                            return result;
+                        })
+                )
+        );
+
+        return configurator.getConfig(BUConfig.class);
     }
 }
